Track the type of benchmark test we're performing
The tests so far have been bare, without any attempt at CPU or memory
interference. We're going to start testing this, so we need to keep
track of the type of test we're performing.

Also realized I don't need to pass in the benchmark address. I use
the same deterministic one every time.